started restructuring DOM and styles
