+++ public/index.html
@@ -27,22 +27,9 @@
       display: none;
     }
 
-    skraafoto-direction-picker {
-      display: block;
-      height: 0;
-      width: 0;
-    }
-
-    @media screen and (max-width: 40rem) {
-
-      .ds-logo-micro>span {
-        display: none;
-      }
-    }
-
     @media screen and (min-width: 80rem) {
 
-      body>main {
+      body > main {
         display: flex;
         flex-flow: row nowrap;
       }
@@ -52,12 +39,6 @@
         width: 70%;
         flex: 1 1 70%;
         position: relative;
-      }
-
-      skraafoto-direction-picker {
-        height: 100%;
-        width: 40rem;
-        flex: 0 0 40rem;
       }
 
     }
