+++ public/singleview.html
@@ -22,13 +22,6 @@
             height: 100%
         }
 
-        @media screen and (max-width: 40rem) {
-
-            .ds-logo-micro > span {
-                display: none;
-            }
-        }
-
         @media screen and (min-width: 80rem) {
 
             body > main {
