+++ public/twinview.html
@@ -25,13 +25,6 @@
       height: 100%;
       width: 100%;
     }
-
-    @media screen and (max-width: 40rem) {
-
-      .ds-logo-micro > span {
-        display: none;
-      }
-    }
   </style>
 </head>
 <body class="sf-app">
