+++ src/components/viewport/viewport.css
@@ -1,0 +1,163 @@
+skraafoto-viewport button {
+  border: none; 
+}
+skraafoto-viewport .basic-image-info {
+  position: absolute;
+  top: var(--space);
+  left: var(--space);
+}
+skraafoto-viewport .sf-viewport-tools {
+  position: absolute;
+  z-index: 2;
+  top: .5rem;
+  left: .5rem;
+  border-radius: 2rem
+}
+skraafoto-viewport .sf-viewport-tools button {
+  display: flex;
+}
+skraafoto-viewport .viewport-map {
+  width: 100%; 
+  height: 100%;
+  position: relative;
+  background-color: var(--background-color);
+}
+skraafoto-viewport .image-date {
+  position: absolute;
+  bottom: 1rem;
+  left: 1rem;
+  color: #fff;
+  margin: 0;
+  -webkit-transform: translate3d(0,0,0); /* Fix for Safari bug */
+}
+skraafoto-viewport skraafoto-compass-arrows {
+  position: absolute;
+  top: var(--space);
+  right: var(--space);
+}
+skraafoto-viewport ds-spinner {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  margin-left: -2rem;
+  margin-top: -2rem;
+  width: 4rem !important;
+  height: 4rem !important;
+  z-index: 10
+}
+skraafoto-viewport ds-spinner > .ds-loading-svg {
+  max-width: 5rem !important;
+  background-color: var(--background-color);
+  border-radius: 50%;
+  padding: 0.75rem;
+}
+skraafoto-viewport .out-of-bounds {
+  display: none;
+  margin: 0;
+  position: absolute;
+  top: 50%;
+  width: 100%;
+  -ms-transform: translateY(-50%);
+  transform: translateY(-50%);
+  text-align: center;
+}
+skraafoto-viewport .ol-viewport canvas {
+  cursor: crosshair;
+}
+skraafoto-viewport .crosshair-on .ol-viewport canvas {
+  cursor: url('../../../public/img/icons/icon_cursor_crosshair.svg') 12 12, crosshair;
+}
+skraafoto-viewport .image-date {
+  display: none;
+}
+skraafoto-viewport .sf-fullscreen-btn {
+  position: absolute;
+  top: 6rem;
+  right: 2rem;
+}
+skraafoto-viewport .sf-fullscreen-btn-true svg:has(use),
+skraafoto-viewport .sf-fullscreen-btn-false svg:has(use) {
+  display: none;
+  margin: 0 !important;
+}
+skraafoto-viewport .sf-fullscreen-btn-true svg.fullscreen-true,
+skraafoto-viewport .sf-fullscreen-btn-false svg.fullscreen-false {
+  display: flex;
+}
+skraafoto-viewport .ol-zoom {
+  bottom: 2rem;
+  right: 1rem;
+  position: absolute;
+}
+skraafoto-viewport .ol-zoom-in,
+skraafoto-viewport .ol-zoom-out {
+  margin: 0.25rem 0 0;
+  display: flex;
+  box-shadow: 0 0.15rem 0.3rem hsl(0, 0%, 50%, 0.5);
+}
+skraafoto-viewport .sf-viewport-tools button.active {
+  background-color: var(--highlight) !important;
+}
+
+/* Info tool, exposure tool */
+skraafoto-viewport .sf-info-btn, 
+skraafoto-viewport .exposure-btn,
+skraafoto-viewport .sf-download-tool,
+skraafoto-viewport .sf-print-tool {
+  border-radius: 0;
+}
+
+/* Measure width tool */
+skraafoto-viewport .sf-tooltip-measure {
+  background-color: var(--background-color);
+  color: var(--color);
+  padding: 0.25rem 0.5rem;
+}
+
+skraafoto-viewport .sf-compass-arrows {
+  display: absolute;
+  padding:10rem;
+}
+
+/* Style last child in btn group */
+skraafoto-viewport .ds-button-group > *:last-child button {
+  border-radius: 0 2.5rem 2.5rem 0;
+  width: 3.5rem !important;
+}
+
+@media screen and (max-width: 35rem) {
+  skraafoto-viewport .sf-fullscreen-btn {
+    top: auto;
+    bottom: 2rem;
+    left: 2rem;
+  }
+  
+  skraafoto-viewport .ol-zoom {
+    display: none;
+    display: none;
+  }
+
+  skraafoto-viewport skraafoto-compass {
+    top: 5.5rem;
+    right: 1.5rem;
+  }
+  skraafoto-viewport skraafoto-compass-arrows {
+    top: auto;
+    bottom: 2rem;
+  }
+  skraafoto-viewport .image-date {
+    bottom: 0.5rem;
+    left: 0.5rem;
+  }
+}
+
+@media screen and (max-width: 50rem) {
+
+  skraafoto-viewport .image-date {
+    display: block;
+    bottom: auto;
+    top: 3.5rem;
+    left: 2rem;
+  }
+
+}
