+++ src/index.css
@@ -2,8 +2,15 @@
 /* CSS imports*/
 @import "@dataforsyningen/designsystem/assets/designsystem.css";
 
-/** Custom styles for cookieinformation popup */
+/* Custom styles for cookieinformation popup */
 @import "./styles/cookieinformation.css";
+
+/* Web component styles */
+@import "./components/viewport/viewport.css";
+@import "./components/viewport/viewport-mini.css";
+@import "./components/viewport/direction-picker.css";
+@import "./components/alert/alert-splash.css";
+@import "./components/compass/compass.css";
 
 /* Tools and buttons */
 @import "./components/tools/measure.css";
